added the full implementation of adding jars and datapacks
