+++ package.json
@@ -39,7 +39,7 @@
     "prettier-format": "prettier --write \"**/*.{ts,tsx,css,yml}\"",
     "setup": "yarn && yarn build",
     "start": "yarn workspace @tsconline/server start",
-    "dev": "yarn dev:config --overwrite && yarn workspace @tsconline/server run migrate-datapacks && concurrently \"yarn start\" \"yarn workspace @tsconline/server dev\" \"wait-on http://localhost:3000 && yarn workspace @tsconline/app dev\" \"yarn workspace @tsconline/shared dev\"",
+    "dev": "yarn workspace @tsconline/server run pull-supporting-files && yarn dev:config --overwrite && yarn workspace @tsconline/server run migrate-datapacks && concurrently \"yarn start\" \"yarn workspace @tsconline/server dev\" \"wait-on http://localhost:3000 && yarn workspace @tsconline/app dev\" \"yarn workspace @tsconline/shared dev\"",
     "production": "yarn workspace @tsconline/server run production",
     "test": "vitest run",
     "test:watch": "vitest",
